+++ packages/react-ui/index.tsx
@@ -1,1 +1,1 @@
-export * from './src';
+export * from './atoms';
