--- packages/react-ui/src/components/index.ts
@@ -1,3 +1,0 @@
-import * as Atoms from './atoms';
-
-export { Atoms };
