--- packages/react-ui/src/index.ts
@@ -1,1 +1,0 @@
-export * from './components';
